Move pin-code template to /stategies
