+++ lib/habilitations/routes.js
@@ -8,7 +8,7 @@ const w = require('../util/w')
 const {createAuthClient} = require('../clients')
 const {createHabilitation, askHabilitation, validateHabilitation, fetchHabilitation} = require('./model')
 const {sendMail} = require('../util/sendmail')
-const createPinCodeEmail = require('../emails/pin-code')
+const createPinCodeEmail = require('./strategies/email/pin-code-template')
 const {pinCodeValidation, hasBeenSentRecently, generatePinCode, getExpirationDate} = require('./strategies/email')
 
 async function habilitationsRoutes(params = {}) {
