+++ lib/habilitations/strategies/email/pin-code-template.js
